Fixed session check, added /channel/users temporary endpoint to get all users, replaced all hardcoded data in ChannelService for database queries
Fix wrong endpoint being visited to check user session.
Added endpoint /channel/users for user search for creation of DMs (issue #45)
Replaced all hardcoded values in the ChannelService to use database queries for our Supabase database (related to US-1 #4 , US-2 #23 , US-3 #36 , US-11 #51 )
Added functionality for sending a message to a channel. (related to US-2 and US-5 #13 )
Added helper methods to execute a query and to execute an update to the database.